created datafile code split
