+++ src/components/ImageMap.jsx
@@ -1,83 +1,20 @@
+// src/components/ImageMap.js
+
 import React, { useEffect } from "react";
 import ClickableArea from "./ClickableArea";
-import shogunGeishaImage from "../../src/assets/images/a Japanese shogun-geisha- red cherry blossom.webp";
 
-const originalImageDimensions = {
-    width: 1792,
-    height: 1024,
-};
-
-// Define clickable areas relative to original image dimensions
-const clickableAreas = {
-    shogun: {
-        x1: 700 / originalImageDimensions.width,
-        y1: 450 / originalImageDimensions.height,
-        x2: 920 / originalImageDimensions.width,
-        y2: 650 / originalImageDimensions.height,
-        label: "Shogun",
-    },
-    geisha: {
-        x1: 1000 / originalImageDimensions.width,
-        y1: 650 / originalImageDimensions.height,
-        x2: 1200 / originalImageDimensions.width,
-        y2: 750 / originalImageDimensions.height,
-        label: "Geisha",
-    },
-    cherry_blossom: {
-        x1: 57 / originalImageDimensions.width,
-        y1: 100 / originalImageDimensions.height,
-        x2: 357 / originalImageDimensions.width,
-        y2: 400 / originalImageDimensions.height,
-        label: "Cherry Blossom",
-    },
-    swordsBelt: {
-        x1: 580 / originalImageDimensions.width,
-        y1: 600 / originalImageDimensions.height,
-        x2: 780 / originalImageDimensions.width,
-        y2: 800 / originalImageDimensions.height,
-        label: "swordsBelt",
-    },
-    kabuto: {
-        x1: 626 / originalImageDimensions.width,
-        y1: 150 / originalImageDimensions.height,
-        x2: 940 / originalImageDimensions.width,
-        y2: 340 / originalImageDimensions.height,
-        label: "kabuto",
-    },
-    kanzashi: {
-        x1: 1005 / originalImageDimensions.width,
-        y1: 250 / originalImageDimensions.height,
-        x2: 1200 / originalImageDimensions.width,
-        y2: 400 / originalImageDimensions.height,
-        label: "kanzashi",
-    },
-    oshiroi: {
-        x1: 1000 / originalImageDimensions.width,
-        y1: 400 / originalImageDimensions.height,
-        x2: 1200 / originalImageDimensions.width,
-        y2: 600 / originalImageDimensions.height,
-        label: "oshiroi",
-    },
-    kimono: {
-        x1: 1000 / originalImageDimensions.width,
-        y1: 750 / originalImageDimensions.height,
-        x2: 1200 / originalImageDimensions.width,
-        y2: 830 / originalImageDimensions.height,
-        label: "kimono",
-    },
-};
-
-const ImageMap = ({ onAreaClick, imageSize, setImageSize, showClickableAreas, modalOpen }) => {
+const ImageMap = ({ image, imageSize, setImageSize, showClickableAreas, onAreaClick }) => {
+    // Auto-adjust image size based on the window's size or image dimensions
     useEffect(() => {
         const updateImageSize = () => {
-            const image = document.getElementById("main-image");
-            if (image) {
-                const { width, height } = image.getBoundingClientRect();
+            const img = document.getElementById("main-image");
+            if (img) {
+                const { width, height } = img.getBoundingClientRect();
                 setImageSize({ width, height });
             }
         };
 
-        // Run on mount and update when window is resized
+        // Run on mount and on resize
         updateImageSize();
         window.addEventListener("resize", updateImageSize);
 
@@ -88,19 +25,19 @@ const ImageMap = ({ onAreaClick, imageSize, setImageSize, showClickableAreas, mo
         <div className="relative mx-auto max-w-full">
             <img
                 id="main-image"
-                src={shogunGeishaImage}
-                alt="Shogun and Geisha Image"
+                src={image.src} // Dynamically use the image source
+                alt={image.alt}
                 className="block max-w-full h-auto"
             />
-            {Object.keys(clickableAreas).map((areaKey) => (
+            {/* Render clickable areas */}
+            {Object.keys(image.clickableAreas).map((areaKey) => (
                 <ClickableArea
                     key={areaKey}
-                    area={clickableAreas[areaKey]}
+                    area={image.clickableAreas[areaKey]}
                     imageSize={imageSize}
                     onClick={() => onAreaClick(areaKey)}
                     showBorders={showClickableAreas}
                     showTooltip={showClickableAreas}
-                    isModalOpen={modalOpen}
                 />
             ))}
         </div>
